requirements.txt
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
email-validator==2.1.1
et-xmlfile==1.1.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Testing==0.8.1
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
pydantic==2.7.0
pydantic-core==2.18.1
pyjwt==2.8.0
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
supabase==2.4.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
webdriver-manager==4.0.1
Werkzeug==3.0.2
wsproto==1.2.0
